+++ src/main/java/com/investmetic/domain/user/dto/response/CustomUserDetails.java
@@ -1,34 +1,26 @@
 package com.investmetic.domain.user.dto.response;
 
-import com.investmetic.domain.user.model.Role;
 import com.investmetic.domain.user.model.entity.User;
-import java.util.ArrayList;
 import java.util.Collection;
+import java.util.Collections;
 import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 
 public class CustomUserDetails implements UserDetails {
 
     private final User user;
-    private final Role role;
 
-    public CustomUserDetails(User user, Role role) {
+    public CustomUserDetails(User user) {
         this.user = user;
-        this.role = role;
     }
 
     @Override
     public Collection<? extends GrantedAuthority> getAuthorities() {
-        Collection<GrantedAuthority> collection = new ArrayList<>();
-
-        collection.add(new GrantedAuthority() {
-            @Override
-            public String getAuthority() {
-                return role.name();
-            }
-        });
-
-        return collection;
+        // User 객체에서 Role을 가져와 GrantedAuthority로 변환
+        return Collections.singletonList(
+                new SimpleGrantedAuthority(user.getRole().name())
+        );
     }
