+++ src/main/java/com/investmetic/global/security/filter/LoginFilter.java
@@ -1,4 +1,4 @@
-package com.investmetic.global.security.jwt;
+package com.investmetic.global.security.filter;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.investmetic.global.util.JWTUtil;
