+++ src/main/java/com/investmetic/global/security/handler/CustomAuthenticationFailureHandler.java
@@ -6,7 +6,6 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.util.Map;
-import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.authentication.AuthenticationFailureHandler;
 import org.springframework.stereotype.Component;
@@ -19,11 +18,7 @@ public class CustomAuthenticationFailureHandler implements AuthenticationFailure
                                         AuthenticationException exception) throws IOException {
 
         ObjectMapper mapper = new ObjectMapper();
-        ErrorCode errorCode = ErrorCode.LOGIN_FAILED;
-
-        if (exception instanceof BadCredentialsException) {
-            errorCode = ErrorCode.LOGIN_FAILED;
-        }
+        ErrorCode errorCode = ErrorCode.LOGIN_FAILED; // 기본 에러코드
 
         // 응답 설정
         response.setStatus(errorCode.getStatus().value());
